# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SIQE Studio API",
    "description": "Enhance, edit, organise and automate images, and train the models that do it.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/assets": {
      "post": {
        "tags": [
          "assets"
        ],
        "summary": "Upload an image (raw request body)",
        "description": "Send the file bytes as the request body and its name in `filename`. The body is streamed to disk and hashed, never held in memory. Uploading a file that is already in the library returns the existing image with `duplicate: true` and status 200.",
        "operationId": "upload_asset_api_assets_post",
        "parameters": [
          {
            "name": "filename",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 255,
              "title": "Filename"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadOut"
                }
              }
            }
          },
          "200": {
            "description": "Already in the library",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/octet-stream": {
              "schema": {}
            }
          }
        }
      },
      "get": {
        "tags": [
          "assets"
        ],
        "summary": "Images in the library, newest first",
        "operationId": "list_assets_api_assets_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 200,
              "title": "Limit"
            }
          },
          {
            "name": "parent_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only images made from this one by AI runs.",
              "title": "Parent Id"
            },
            "description": "Only images made from this one by AI runs."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AssetOut"
                  },
                  "title": "Response List Assets Api Assets Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/assets/{asset_id}": {
      "get": {
        "tags": [
          "assets"
        ],
        "summary": "One image",
        "operationId": "one_asset_api_assets__asset_id__get",
        "parameters": [
          {
            "name": "asset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Asset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "assets"
        ],
        "summary": "Delete an image and its files",
        "operationId": "delete_asset_api_assets__asset_id__delete",
        "parameters": [
          {
            "name": "asset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Asset Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/assets/{asset_id}/thumb": {
      "get": {
        "tags": [
          "assets"
        ],
        "summary": "320 px thumbnail (WebP)",
        "operationId": "asset_thumb_api_assets__asset_id__thumb_get",
        "parameters": [
          {
            "name": "asset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Asset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/assets/{asset_id}/preview": {
      "get": {
        "tags": [
          "assets"
        ],
        "summary": "2048 px preview (WebP)",
        "operationId": "asset_preview_api_assets__asset_id__preview_get",
        "parameters": [
          {
            "name": "asset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Asset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/assets/{asset_id}/dz/{path}": {
      "get": {
        "tags": [
          "assets"
        ],
        "summary": "Deep-zoom pyramid files",
        "operationId": "asset_deep_zoom_api_assets__asset_id__dz__path__get",
        "parameters": [
          {
            "name": "asset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Asset Id"
            }
          },
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/assets/{asset_id}/original": {
      "get": {
        "tags": [
          "assets"
        ],
        "summary": "Download the original file",
        "operationId": "asset_original_api_assets__asset_id__original_get",
        "parameters": [
          {
            "name": "asset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Asset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/assets/{asset_id}/edits": {
      "get": {
        "tags": [
          "assets"
        ],
        "summary": "The image's edit document",
        "operationId": "get_edits_api_assets__asset_id__edits_get",
        "parameters": [
          {
            "name": "asset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Asset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditDocumentIn"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "assets"
        ],
        "summary": "Replace the image's edit document",
        "operationId": "put_edits_api_assets__asset_id__edits_put",
        "parameters": [
          {
            "name": "asset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Asset Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditDocumentIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditDocumentIn"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/assets/{asset_id}/exports": {
      "post": {
        "tags": [
          "assets"
        ],
        "summary": "Render the current edits at full resolution and encode them",
        "operationId": "start_export_api_assets__asset_id__exports_post",
        "parameters": [
          {
            "name": "asset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Asset Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportStartOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/assets/{asset_id}/renditions": {
      "get": {
        "tags": [
          "assets"
        ],
        "summary": "Exports of this image",
        "operationId": "list_renditions_api_assets__asset_id__renditions_get",
        "parameters": [
          {
            "name": "asset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Asset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RenditionOut"
                  },
                  "title": "Response List Renditions Api Assets  Asset Id  Renditions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/renditions/{rendition_id}/download": {
      "get": {
        "tags": [
          "assets"
        ],
        "summary": "Download an export",
        "operationId": "download_rendition_api_renditions__rendition_id__download_get",
        "parameters": [
          {
            "name": "rendition_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Rendition Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/renditions/{rendition_id}": {
      "delete": {
        "tags": [
          "assets"
        ],
        "summary": "Delete an export",
        "operationId": "delete_rendition_api_renditions__rendition_id__delete",
        "parameters": [
          {
            "name": "rendition_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Rendition Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EditDocumentIn": {
        "properties": {
          "version": {
            "type": "integer",
            "const": 1,
            "title": "Version",
            "default": 1
          },
          "geometry": {
            "$ref": "#/components/schemas/GeometryIn"
          },
          "ops": {
            "items": {
              "$ref": "#/components/schemas/OpEntryIn"
            },
            "type": "array",
            "title": "Ops"
          }
        },
        "type": "object",
        "title": "EditDocumentIn",
        "description": "Validated against the operation catalog on save; see GET /api/ops."
      },
      "ExportIn": {
        "properties": {
          "format": {
            "type": "string",
            "enum": [
              "jpeg",
              "png",
              "webp",
              "avif",
              "tiff"
            ],
            "title": "Format",
            "default": "jpeg"
          },
          "quality": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Quality",
            "default": 90
          },
          "max_side": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 200000.0,
                "minimum": 16.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Side"
          },
          "target_kb": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 2000000.0,
                "minimum": 10.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Kb"
          },
          "strip_metadata": {
            "type": "boolean",
            "title": "Strip Metadata",
            "default": true
          }
        },
        "type": "object",
        "title": "ExportIn"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UploadOut": {
        "properties": {
          "asset": {
            "$ref": "#/components/schemas/AssetOut"
          },
          "duplicate": {
            "type": "boolean",
            "title": "Duplicate",
            "description": "True when this exact file was already in the library."
          },
          "job": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "asset",
          "duplicate",
          "job"
        ],
        "title": "UploadOut"
      },
...
    }
  }
}